docs: remove app instance from api.py
Detail:

While i'm developing and test the docs of this app in development env,
i found that i cannot to this if i import the orders/api/api module
directly to the orders/app module. So, for fix that in temporary time,
i create the instance inside orders/api/api. Now, I've override the
docs with my own specification in oas.yaml that I'll push it soon.